* refactor: move message in ValidationError * fix: exception in BannedUserVerify * refactor: add return value in method _condition
