+++ backend/users/backend/base/validators.py
@@ -1,31 +1,30 @@
 from abc import ABCMeta
-from typing import Union, Type
+from typing import Type
 
-from django.utils.translation import gettext_lazy as _
-
-from administrator.models import Admin
-from customer.models import CustomerUser
-from developer.models import CompanyUser
+from django.contrib.auth.models import AbstractUser
 
 
 class AbstractUserVerification(metaclass=ABCMeta):
     """ Common abstract class for user verification """
 
-    def verify(self, user: Union[CustomerUser, Admin, CompanyUser]):
+    def verify(self, user: AbstractUser):
         pass
 
 
 class VerificationError(Exception):
     """ Common base class for all verification exceptions. """
+    message = 'user verification error'
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(self.message, *args)
 
 
 class BaseUserVerification(AbstractUserVerification):
     exception: Type[VerificationError]
-    message: str
 
-    def _condition(self, user: Union[CustomerUser, Admin, CompanyUser]):
+    def _condition(self, user) -> bool:
         raise NotImplementedError
 
     def verify(self, user):
         if self._condition(user):
-            raise self.exception(_(self.message))
+            raise self.exception()
